+++ serde_arrow/src/arrow2_impl/schema.rs
@@ -11,39 +11,6 @@ use crate::{
     },
 };
 
-/// Support for arrow2 types (*requires one of the `arrow2-*` features*)
-impl SerdeArrowSchema {
-    /// Build a new Schema object from fields
-    pub fn from_arrow2_fields(fields: &[Field]) -> Result<Self> {
-        Self::try_from(fields)
-    }
-
-    /// This method is deprecated. Use
-    /// [`to_arrow2_fields`][SerdeArrowSchema::to_arrow2_fields] instead:
-    ///
-    /// ```rust
-    /// # fn main() -> serde_arrow::_impl::PanicOnError<()> {
-    /// # use serde_arrow::schema::{SerdeArrowSchema, SchemaLike, TracingOptions};
-    /// # #[derive(serde::Deserialize)]
-    /// # struct Item { a: u32 }
-    /// # let schema = SerdeArrowSchema::from_type::<Item>(TracingOptions::default()).unwrap();
-    /// # let fields =
-    /// schema.to_arrow2_fields()?
-    /// # ;
-    /// # Ok(())
-    /// # }
-    /// ```
-    #[deprecated = "The method `get_arrow2_fields` is deprecated. Use `to_arrow2_fields` instead"]
-    pub fn get_arrow2_fields(&self) -> Result<Vec<Field>> {
-        Vec::<Field>::try_from(self)
-    }
-
-    /// Build a vec of fields from a  Schema object
-    pub fn to_arrow2_fields(&self) -> Result<Vec<Field>> {
-        Vec::<Field>::try_from(self)
-    }
-}
-
 impl TryFrom<SerdeArrowSchema> for Vec<Field> {
     type Error = Error;
 
@@ -79,20 +46,20 @@ impl Sealed for Vec<Field> {}
 /// `arrow2-*` features*)
 impl SchemaLike for Vec<Field> {
     fn from_value<T: serde::Serialize + ?Sized>(value: &T) -> Result<Self> {
-        SerdeArrowSchema::from_value(value)?.to_arrow2_fields()
+        SerdeArrowSchema::from_value(value)?.try_into()
     }
 
     fn from_type<'de, T: serde::Deserialize<'de> + ?Sized>(
         options: crate::schema::TracingOptions,
     ) -> Result<Self> {
-        SerdeArrowSchema::from_type::<T>(options)?.to_arrow2_fields()
+        SerdeArrowSchema::from_type::<T>(options)?.try_into()
     }
 
     fn from_samples<T: serde::Serialize + ?Sized>(
         samples: &T,
         options: crate::schema::TracingOptions,
     ) -> Result<Self> {
-        SerdeArrowSchema::from_samples(samples, options)?.to_arrow2_fields()
+        SerdeArrowSchema::from_samples(samples, options)?.try_into()
     }
 }