+++ serde_arrow/src/lib.rs
@@ -325,19 +325,11 @@ mod arrow_impl;
 #[cfg(has_arrow)]
 pub use arrow_impl::api::{from_arrow, from_record_batch, to_arrow, to_record_batch};
 
-#[cfg(has_arrow)]
-#[allow(deprecated)]
-pub use arrow_impl::api::ArrowBuilder;
-
 #[cfg(has_arrow2)]
 mod arrow2_impl;
 
 #[cfg(has_arrow2)]
 pub use arrow2_impl::api::{from_arrow2, to_arrow2};
-
-#[cfg(has_arrow2)]
-#[allow(deprecated)]
-pub use arrow2_impl::api::Arrow2Builder;
 
 #[deny(missing_docs)]
 /// Helpers that may be useful when using `serde_arrow`
@@ -394,10 +386,6 @@ pub mod schema {
     pub use crate::internal::schema::{
         Overwrites, SchemaLike, SerdeArrowSchema, Strategy, TracingOptions, STRATEGY_KEY,
     };
-
-    /// Renamed to [`SerdeArrowSchema`]
-    #[deprecated = "serde_arrow::schema::Schema is deprecated. Use serde_arrow::schema::SerdeArrowSchema instead"]
-    pub type Schema = SerdeArrowSchema;
 
     /// Support for [canonical extension types][ext-docs]. This module is experimental without semver guarantees.
     ///