+++ serde_arrow/src/test_with_arrow/impls/utils.rs
@@ -16,7 +16,7 @@ pub struct Arrays {
 
 #[derive(Default)]
 pub struct Fields {
-    pub arrow: Option<Vec<arrow::datatypes::Field>>,
+    pub arrow: Option<Vec<arrow::datatypes::FieldRef>>,
     pub arrow2: Option<Vec<arrow2::datatypes::Field>>,
 }
 
@@ -77,11 +77,10 @@ impl Test {
 }
 
 impl Test {
-    pub fn get_arrow_fields(&self) -> Cow<'_, Vec<arrow::datatypes::Field>> {
+    pub fn get_arrow_fields(&self) -> Cow<'_, Vec<arrow::datatypes::FieldRef>> {
         match self.schema.as_ref() {
             Some(schema) => Cow::Owned(
-                schema
-                    .to_arrow_fields()
+                Vec::<arrow::datatypes::FieldRef>::try_from(schema)
                     .expect("Cannot covert schema to arrow fields"),
             ),
             None => Cow::Borrowed(
@@ -96,8 +95,7 @@ impl Test {
     pub fn get_arrow2_fields(&self) -> Cow<'_, Vec<arrow2::datatypes::Field>> {
         match self.schema.as_ref() {
             Some(schema) => Cow::Owned(
-                schema
-                    .to_arrow2_fields()
+                Vec::<arrow2::datatypes::Field>::try_from(schema)
                     .expect("Cannot covert schema to arrow fields"),
             ),
             None => Cow::Borrowed(
@@ -146,10 +144,6 @@ impl Test {
 
     pub fn try_serialize_arrow<T: Serialize + ?Sized>(&mut self, items: &T) -> Result<()> {
         let fields = self.get_arrow_fields().to_vec();
-        let field_refs = fields
-            .iter()
-            .map(|f| Arc::new(f.clone()))
-            .collect::<Vec<_>>();
         let arrays = crate::to_arrow(&fields, items)?;
 
         assert_eq!(fields.len(), arrays.len());
@@ -168,7 +162,7 @@ impl Test {
 
         self.arrays.arrow = Some(arrays);
 
-        let mut builder = crate::ArrayBuilder::from_arrow(&field_refs)?;
+        let mut builder = crate::ArrayBuilder::from_arrow(&fields)?;
         builder.extend(items)?;
         let arrays = builder.to_arrow()?;
         assert_eq!(self.arrays.arrow, Some(arrays));